fix dashboard time
